Changed consturcotr for player object from int to floats
Initially was ints in testing but when extending actor needs to become float to prevent conversion errros.

+++ PotassiumOverload/core/src/com/cucubananas/core/actor/AnimatedMoveableObject.java
@@ -6,7 +6,7 @@ import java.util.HashMap;
 public class AnimatedMoveableObject extends MoveableObject {
   HashMap<String, TextureRegion> spriteSheet;
 
-  public AnimatedMoveableObject(String texturePath, int xPos, int yPos) {
+  public AnimatedMoveableObject(String texturePath, float xPos, float yPos) {
     super(texturePath, xPos, yPos);
     spriteSheet = new HashMap<>();
   }
